% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{custom_palettes}
\alias{custom_palettes}
\title{Custom Palettes}
\format{
\code{custom_palettes}
}
\source{
Made for pacakge
}
\usage{
custom_palettes
}
\description{
Custom palettes to use for the trout.modlR package
}
\keyword{datasets}
