% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{format_axis}
\alias{format_axis}
\title{Format Axis}
\usage{
format_axis(form)
}
\arguments{
\item{form}{A numeric object specifying the axis them (1 or 2)}
}
\value{
A ggplot2 theme
}
\description{
A helper function used to build a custom theme for the axes.
}
\examples{
ggplot2::ggplot() + format_axis(1)
}
