% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{format_legend}
\alias{format_legend}
\title{Format Legend}
\usage{
format_legend(side)
}
\arguments{
\item{side}{A character object specifying the in which legend will be
positioned ("top" or "side").}
}
\value{
A ggplot2 theme
}
\description{
A helper function used to format a ggplot2 graph's legend.
}
\examples{
ggplot2::ggplot() + format_legend("top")
}
