% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomass_data.R
\name{get_biomass_data}
\alias{get_biomass_data}
\title{Get Biomass Data}
\usage{
get_biomass_data(
  dat1,
  dat2 = NULL,
  dat3 = NULL,
  grouping = c("year", "data_type"),
  cumulative = FALSE,
  pounds = TRUE
)
}
\arguments{
\item{dat1}{An rdat file containing a model output}

\item{dat2}{An rdat file containing a model output}

\item{dat3}{An rdat file containing a model output}

\item{grouping}{Grouping parameters. This determines which groups will be
used to sum biomass if cumulative = TRUE. Options include \code{year},
\code{data_type}, \code{mu}, \code{species}, and \code{model}.}

\item{cumulative}{TRUE of FALSE(default).  If TRUE, the sum of biomass will
be calculate by \code{grouping}. This can work when multiple models are supplied}

\item{pounds}{TRUE or FALSE.  If TRUE, converts from kg's to pounds.}

\item{wide}{TRUE or FALSE. If TRUE, output the data in wide format}
}
\value{
A tibble
}
\description{
\code{get_biomass_data} can be used to extract biomass data from .rdat
model outputs. This can be used to extract data from a single model, multiple
models, or calculate the sum of biomass by grouping parameters.
}
\examples{
get_biomass_data(trout1)
get_biomass_data(trout1,trout2,grouping = c("data_type","mu"), cumulative = TRUE)
}
