% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameter}
\alias{get_parameter}
\title{Get Parameter}
\usage{
get_parameter(dat, param_name)
}
\arguments{
\item{dat}{An .rdat file}

\item{param_name}{A character object specifying a parameter name. See
\code{parameter_names} for details.}
}
\value{
Data specified by \code{param_nam}.  This can be a list, a data.frame,
an integer, or other data types.
}
\description{
\code{get_parameter} parses .rdat data output by an ADMB model.
\code{get_parameter} can only retrieve a single parameter at a time. This
function is used within \code{get_parameters} to retrieve multiple parameters.
Available options for \code{param_name} can be found in \code{parameter_names}.
}
\examples{
get_parameter(trout1,"years")

}
