% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{get_parameters}
\alias{get_parameters}
\title{Get model Parameters}
\usage{
get_parameters(dat, param_names)
}
\arguments{
\item{dat}{A list}

\item{param_name}{A parameter name (s)}
}
\value{
A vector
}
\description{
These functions takes an .rdat file output by ADMB and a param_name. The function
will return data for the given parameter.  See parameter_lookup to see
available options.
}
\examples{
get_parameters(gb_wf_ar,"years")
}
