% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Nplots.R
\name{graph_Nplot}
\alias{graph_Nplot}
\title{Graph Nplot}
\usage{
graph_Nplot(data, type, title = "Y", ymax = 0)
}
\arguments{
\item{type}{Graph Type("total","recruit",or "naa")}

\item{title}{Include graph title ("Y" or "N")}

\item{ymax}{Y Axis Maximum}

\item{dat}{ADMB data output}
}
\value{
A ggplot object
}
\description{
A wrapper function used to create Nplot. The type arguement determines
which Nplot graphing function will be used ("total","recruit",or "naa").
}
\examples{
graph_nplot()
}
