% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bmsPlots.R
\name{make_bmsPlot}
\alias{make_bmsPlot}
\title{Plot estimated biomass}
\usage{
make_bmsPlot(dat, title = NULL, ymax = NULL, plotly_plot = FALSE)
}
\arguments{
\item{dat}{A data.frame}

\item{title}{A plot title.  If NULL, a default title will be used}

\item{ymax}{A numeric object specifying the maximum value of the y-axis.
If NULL, the value will be auto selected.}

\item{plotly_plot}{TRUE or FALSE.  If TRUE, a \code{plotly::ggplotly()}
object will be output. If FALSE (default), a \code{ggplot2::ggplot()}
object will be output instead.}
}
\value{
A \code{ggplot2::ggplot()} or \code{plotly::ggplotly()} object
}
\description{
This function plots the estimated biomass.
}
\examples{
make_bmsPlot(dat = gb_wf_ar)
}
