% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameter_names}
\alias{parameter_names}
\title{Parameter Lookup Table}
\format{
\code{parameter_names} is a data.frame with 24 rows and 2 variables as
follows:
\describe{
  \item{param_name}{Name of parameter. This can be supplied to
        \code{get_parameter} and \code{get_parameter}}
  \item{descriptions}{A text description of the parameters}
}
}
\source{
parameter_names-- Built by hand for trout.modlR
}
\usage{
parameter_names
}
\description{
This table lists all of the available options for the \code{get_parameter}
and \code{get_parameter} functions. A description is provided for each
option.
}
\keyword{datasets}
