% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass_data.R
\name{plot_biomass_data}
\alias{plot_biomass_data}
\title{Plot Biomass Data}
\usage{
plot_biomass_data(
  dat,
  title = NULL,
  ymax = NULL,
  cumulative = FALSE,
  plotly = FALSE
)
}
\arguments{
\item{title}{A character object containing the plot title.
If left NULL (default), a title will be automatically generated}

\item{ymax}{A number representing the max for the y-axis. If left
null(default), this will be chosen automatically.}

\item{plotly}{TRUE or FALSE (default). If TRUE, a plotly::ggplotly()
object will be returned.}

\item{dat1}{A .rdat containing ADMB output}

\item{dat2}{A .rdat containing ADMB output}

\item{dat3}{A .rdat containing ADMB output}

\item{type}{"single","multiple", or "cumulative". This determines
the type of plot that will be produced. If "single" is selected
model results will be plotted for an individual model.
Only supply a '.rdat' object to 'dat1' when using 'single'.
The 'multiple' argument will add biomass data from multiple
models to one plot. The 'cumulative' type will produce a stacked
area plot containing biomass from multiple models.
When 'cumulative' or 'multiple' are specified, pass additional
'.rdat' files to 'dat2' and/or 'dat3'}
}
\value{
A ggplot2::ggplot() or plotly::ggplotly() object
}
\description{
\code{plot_biomass_data} plots biomass for a given model or models.
 Different graphs can be output by using the 'type' argument.
}
\examples{
plot_biomass_data(dat1 = trout1,type = 'single')
plot_biomass_data(dat1 = trout1,dat2 = trout2, type = 'multiple')
plot_biomass_data(dat1 = trout1,dat2 = trout2, type = 'cumulative')
}
