% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model.R
\name{read_model}
\alias{read_model}
\title{Read Model}
\usage{
read_model(rdat)
}
\arguments{
\item{rdat}{The filepathway to the rdat file}
}
\value{
A list
}
\description{
read_model imports an rdat model file that was exported from ADMB.
}
\examples{
\dontrun{"trout.rdat"}
}
