% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{scale_fill_custom}
\alias{scale_fill_custom}
\title{Apply custom fill palettes to ggplot2 plots}
\usage{
scale_fill_custom(pal)
}
\arguments{
\item{pal}{A character object}
}
\value{

}
\description{
A function used to apply custom color palettes to ggplot2 plots.
See custom_palletes for a list of palette options.  This  was adapted from a
function found in the ggtheme package.
}
\examples{
\dontrun{scale_color_custom("muddy_river")}
}
