% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trout1}
\alias{trout1}
\title{Lake Trout Model output1}
\format{
\code{trout1} is a list with 44 elements
}
\source{
ADMB model output
}
\usage{
trout1
}
\description{
Lake trout data output from an ADMB model. Used within the package as
example data for testing.
}
\keyword{datasets}
