% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trout2}
\alias{trout2}
\title{Lake Trout Model output2}
\format{
\code{trout2} is a list with 44 elements
}
\source{
ADMB model output
}
\usage{
trout2
}
\description{
Lake trout data output from an ADMB model. Used within the package as
example data for testing.
}
\keyword{datasets}
