% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whitefish1}
\alias{whitefish1}
\title{Whitefish Model Output1}
\format{
\code{whitefish1} is a list with 56 elements
}
\source{
ADMB model output
}
\usage{
whitefish1
}
\description{
Whitefish data output from an ADMB model. Used within the package as
example data for testing.
}
\keyword{datasets}
