% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whitefish2}
\alias{whitefish2}
\title{Whitefish Model Output2}
\format{
\code{whitefish2} is a list with 56 elements
}
\source{
ADMB model output
}
\usage{
whitefish2
}
\description{
Lake trout data output from an ADMB model. Used within the package as
example data for testing.
}
\keyword{datasets}
