% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{get_parameters}
\alias{get_parameters}
\title{Get model Parameters}
\usage{
get_parameters(dat, param_names)
}
\arguments{
\item{dat}{A list}

\item{param_name}{A parameter name (s)}
}
\value{
A list
}
\description{
\code{get_parameters} parses .rdat data output by an ADMB model.  The function
accepts a vector of parameter names, parses the data, and outputs a list.
Available optiuons for \code{param_name} can be found in
\code{parameter_names}.
}
\examples{
get_parameters(trout1,c("years","ages"))
}
